requirements.txt
rumps>=0.4.0
pynput>=1.7.6
pyobjc-framework-Quartz>=10.0